+++ src/bundlers/node/nodeJsLocalBundler.ts
@@ -34,38 +34,6 @@ server.listen(port, () => {
 });
 `;
 
-export function getLocalFunctionWrapperCode(handler: string, entry: string) {
-    return `import { ${handler} as userHandler } from "./${entry}";
-
-import http from "http";
-
-const port = process.argv[2];
-
-const server = http.createServer((req, res) => {
-  if (req.method === 'POST') {
-    let body = '';
-    req.on('data', chunk => {
-      body += chunk.toString();
-    });
-    req.on('end', () => {
-      res.writeHead(200, {'Content-Type': 'text/plain'});
-      const jsonParsedBody = JSON.parse(body);
-
-      userHandler(jsonParsedBody).then((response) => {
-        res.end(JSON.stringify(response));
-    })
-    });
-  } else {
-    res.writeHead(404, {'Content-Type': 'text/plain'});
-    res.end('404 Not Found');
-  }
-});
-
-server.listen(port, () => {
-});
-`;
-}
-
 const localClusterWrapperCode = `
 ${localWrapperCode}
 export { server }
